Update orquestra example.
