+++ examples/ex_5_orquestra/run.py
@@ -12,13 +12,12 @@ import time
 
 from orquestra.sdk.schema.workflow_run import State
 
-from .defs import hydrogen_workflow
-from .estimates_ms import ms_estimates
+from .defs import example_workflow
 
 
 def main():
     # uncomment which workflow do you want to run:
-    wf = hydrogen_workflow()  # general hydrogen workflow to calculate estimates
+    wf = example_workflow()  # general hydrogen workflow to calculate estimates
     # wf = ms_estimates()  # calculate resource estimates using MS tool
 
     # Run locally, sequentially, in a single Python process. Useful for debugging, but
@@ -26,7 +25,7 @@ def main():
     wf_run = wf.run("in_process")
 
     # Run remotely on Orquestra Platform.
-    # wf_run = wf.run("prod-d")
+    # wf_run = wf.run("darpa-benchmarking")
 
     print(f"Workflow {wf_run.run_id} submitted!")
